requirements.txt
aiogram==3.27.0.post0
python-dotenv
